chore: relocated issue-related models to separate file
