+++ tests/test_model.py
@@ -3,8 +3,9 @@ from unittest import TestCase
 
 from cyclonedx.exception.model import InvalidLocaleTypeException, InvalidUriException, UnknownHashTypeException
 
-from cyclonedx.model import Encoding, ExternalReference, ExternalReferenceType, HashAlgorithm, HashType, \
-    IssueClassification, IssueType, Note, NoteText, XsUri
+from cyclonedx.model import Encoding, ExternalReference, ExternalReferenceType, HashAlgorithm, HashType, Note, \
+    NoteText, XsUri
+from cyclonedx.model.issue import IssueClassification, IssueType
 
 
 class TestModelExternalReference(TestCase):
